refactor(dashboard): extract JSX for SummaryDisplay component
Extracted the results display section from dashboard page into a dedicated SummaryDisplay component.
Implemented:
- Created types for AISummary, CommitSummary and related interfaces
- Implemented Props interface with all required properties (summary, activityMode, dateRange, activeFilters, installationIds)
- Moved all JSX for the summary display section including ActivityFeed, stats dashboard, and AI summary
- Updated dashboard page to use the new component
