+++ TODO.md
@@ -85,7 +85,7 @@
   - **AC Ref:** 4.1, 4.2, 4.3, 4.4
 
 ## Extract SummaryDisplay Component
-- [ ] **Extract JSX for SummaryDisplay:** Extract results section from dashboard page
+- [x] **Extract JSX for SummaryDisplay:** Extract results section from dashboard page
   - **Action:** Identify and cut the JSX for the results display area (ActivityFeed, stats, AI summary sections) from `src/app/dashboard/page.tsx`, and paste it into `SummaryDisplay.tsx`, ensuring all styles are included.
   - **Depends On:** Create Basic Component Files
   - **AC Ref:** 3.1, 3.2, 3.3
