+++ extract-jsx-for-summarydisplay-PLAN.md
@@ -1,0 +1,4 @@
+# Extract JSX for SummaryDisplay
+
+## Implementation Approach
+Extract the JSX for the summary display section (lines ~705-984) from the dashboard page and move it to the SummaryDisplay component, including the header, ActivityFeed, stats dashboard, and AI summary sections. Define the necessary props interface and types.
